requirements.txt
streamlit
python-dotenv
agno
openai
yfinance
requests
google-generativeai